+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Next
 - Divided js and css resources in two parts (frontend and backend), and extracted vendor files.
+- Can filter region listing by categories.
 
 # 1.4.1 (2020-03-21)
 - Moved to a Github organization for publicly announced open sourcing of the project, so we cleaned up our readme, changelog and contributing guildelines.