+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# Next
+# 1.5.0 (2020-03-22)
 - Divided js and css resources in two parts (frontend and backend), and extracted vendor files.
 - Can filter region listing by categories.